Extract the Pending helper from BatchBuilder
